+++ c/sedona-s2geography/Cargo.toml
@@ -36,7 +36,7 @@ regex = { workspace = true }
 criterion = { workspace = true }
 rstest = { workspace = true }
 sedona = { workspace = true }
-sedona-testing = { workspace = true }
+sedona-testing = { workspace = true, features = ["criterion"] }
 
 [dependencies]
 arrow-schema = { workspace = true }
